refactor: move some tests out of codec_chain.rs, don't re-export the new chunk grid types
