+++ zarrs/src/array/codec/array_to_bytes/codec_chain.rs
@@ -1004,7 +1004,6 @@ mod tests {
     use std::num::NonZeroU64;
 
     use super::*;
-    use crate::array::chunk_grid::RegularChunkGrid;
     use crate::array::codec::BytesCodec;
     use crate::array::{ArraySubset, ArraySubsetTraits, ChunkShapeTraits, data_type};
 
@@ -1165,312 +1164,6 @@ mod tests {
             }
             Ok(input_handle)
         }
-    }
-
-    #[derive(Debug)]
-    struct LocalOnlyReshapeGridCodec {
-        data_type: DataType,
-        fill_value: FillValue,
-        reject_chunk_local: bool,
-    }
-
-    impl ExtensionName for LocalOnlyReshapeGridCodec {
-        fn name(&self, _version: ZarrVersion) -> Option<Cow<'static, str>> {
-            None
-        }
-    }
-
-    impl CodecTraits for LocalOnlyReshapeGridCodec {
-        fn configuration(
-            &self,
-            _version: ZarrVersion,
-            _options: &CodecMetadataOptions,
-        ) -> Option<Configuration> {
-            None
-        }
-
-        fn partial_decoder_capability(&self) -> PartialDecoderCapability {
-            PartialDecoderCapability {
-                partial_read: true,
-                partial_decode: true,
-            }
-        }
-
-        fn partial_encoder_capability(&self) -> PartialEncoderCapability {
-            PartialEncoderCapability {
-                partial_encode: true,
-            }
-        }
-    }
-
-    impl ArrayCodecTraits for LocalOnlyReshapeGridCodec {
-        fn as_any(&self) -> &dyn std::any::Any {
-            self
-        }
-
-        fn data_type(&self) -> &DataType {
-            &self.data_type
-        }
-
-        fn fill_value(&self) -> &FillValue {
-            &self.fill_value
-        }
-
-        fn recommended_concurrency(
-            &self,
-            _shape: &[NonZeroU64],
-        ) -> Result<RecommendedConcurrency, CodecError> {
-            Ok(RecommendedConcurrency::new_maximum(1))
-        }
-    }
-
-    impl ArrayToArrayCodecTraits for LocalOnlyReshapeGridCodec {
-        fn into_dyn(self: Arc<Self>) -> Arc<dyn ArrayToArrayCodecTraits> {
-            self
-        }
-
-        fn encoded_data_type(&self) -> &DataType {
-            &self.data_type
-        }
-
-        fn encoded_fill_value(&self) -> &FillValue {
-            &self.fill_value
-        }
-
-        fn encoded_shape(&self, decoded_shape: &[NonZeroU64]) -> Result<ChunkShape, CodecError> {
-            Ok(vec![
-                NonZeroU64::new(decoded_shape.num_elements_u64()).unwrap(),
-            ])
-        }
-
-        fn encoded_chunk_grid(
-            &self,
-            decoded_chunk_grid: ChunkGridDecodedRef<'_>,
-        ) -> Result<ChunkGridEncoded, ChunkGridCreateError> {
-            match decoded_chunk_grid {
-                ChunkGridDecodedRef::None => Ok(ChunkGridEncoded::None),
-                ChunkGridDecodedRef::Array(_) => Ok(ChunkGridEncoded::ChunkLocal),
-                ChunkGridDecodedRef::ChunkLocal if self.reject_chunk_local => {
-                    Ok(ChunkGridEncoded::None)
-                }
-                ChunkGridDecodedRef::ChunkLocal => Ok(ChunkGridEncoded::ChunkLocal),
-            }
-        }
-
-        fn decoded_subchunk_grid(
-            &self,
-            _decoded_chunk_grid: ChunkGridDecodedRef<'_>,
-            encoded_subchunk_grid: ChunkGridEncodedRef<'_>,
-        ) -> Result<ChunkGridDecoded, ChunkGridCreateError> {
-            Ok(encoded_subchunk_grid.into())
-        }
-
-        fn encode<'a>(
-            &self,
-            _bytes: ArrayBytes<'a>,
-            _shape: &[NonZeroU64],
-            _options: &CodecOptions,
-        ) -> Result<ArrayBytes<'a>, CodecError> {
-            unimplemented!("test codec only exercises subchunk-grid propagation")
-        }
-
-        fn decode<'a>(
-            &self,
-            _bytes: ArrayBytes<'a>,
-            _shape: &[NonZeroU64],
-            _options: &CodecOptions,
-        ) -> Result<ArrayBytes<'a>, CodecError> {
-            unimplemented!("test codec only exercises subchunk-grid propagation")
-        }
-    }
-
-    #[derive(Debug)]
-    struct TestSubchunkingCodec {
-        data_type: DataType,
-        fill_value: FillValue,
-        expose_subchunks: bool,
-    }
-
-    impl ExtensionName for TestSubchunkingCodec {
-        fn name(&self, _version: ZarrVersion) -> Option<Cow<'static, str>> {
-            None
-        }
-    }
-
-    impl CodecTraits for TestSubchunkingCodec {
-        fn configuration(
-            &self,
-            _version: ZarrVersion,
-            _options: &CodecMetadataOptions,
-        ) -> Option<Configuration> {
-            None
-        }
-
-        fn partial_decoder_capability(&self) -> PartialDecoderCapability {
-            PartialDecoderCapability {
-                partial_read: true,
-                partial_decode: true,
-            }
-        }
-
-        fn partial_encoder_capability(&self) -> PartialEncoderCapability {
-            PartialEncoderCapability {
-                partial_encode: true,
-            }
-        }
-    }
-
-    impl ArrayCodecTraits for TestSubchunkingCodec {
-        fn as_any(&self) -> &dyn std::any::Any {
-            self
-        }
-
-        fn data_type(&self) -> &DataType {
-            &self.data_type
-        }
-
-        fn fill_value(&self) -> &FillValue {
-            &self.fill_value
-        }
-
-        fn recommended_concurrency(
-            &self,
-            _shape: &[NonZeroU64],
-        ) -> Result<RecommendedConcurrency, CodecError> {
-            Ok(RecommendedConcurrency::new_maximum(1))
-        }
-    }
-
-    impl ArrayToBytesCodecTraits for TestSubchunkingCodec {
-        fn into_dyn(self: Arc<Self>) -> Arc<dyn ArrayToBytesCodecTraits> {
-            self
-        }
-
-        fn encoded_representation(
-            &self,
-            _shape: &[NonZeroU64],
-        ) -> Result<BytesRepresentation, CodecError> {
-            Ok(BytesRepresentation::UnboundedSize)
-        }
-
-        fn decoded_subchunk_grid(
-            &self,
-            decoded_chunk_grid: ChunkGridDecodedRef<'_>,
-        ) -> Result<ChunkGridDecoded, ChunkGridCreateError> {
-            if !self.expose_subchunks {
-                return Ok(ChunkGridDecoded::None);
-            }
-            match decoded_chunk_grid {
-                ChunkGridDecodedRef::None => Ok(ChunkGridDecoded::None),
-                ChunkGridDecodedRef::ChunkLocal => Ok(ChunkGridDecoded::ChunkLocal),
-                ChunkGridDecodedRef::Array(decoded_chunk_grid) => Ok(ChunkGridDecoded::Array(
-                    ChunkGrid::new(RegularChunkGrid::new(
-                        decoded_chunk_grid.array_shape().to_vec(),
-                        vec![NonZeroU64::new(3).unwrap()],
-                    )?),
-                )),
-            }
-        }
-
-        fn encode<'a>(
-            &self,
-            _bytes: ArrayBytes<'a>,
-            _shape: &[NonZeroU64],
-            _options: &CodecOptions,
-        ) -> Result<ArrayBytesRaw<'a>, CodecError> {
-            unimplemented!("test codec only exercises subchunk-grid propagation")
-        }
-
-        fn decode<'a>(
-            &self,
-            _bytes: ArrayBytesRaw<'a>,
-            _shape: &[NonZeroU64],
-            _options: &CodecOptions,
-        ) -> Result<ArrayBytes<'a>, CodecError> {
-            unimplemented!("test codec only exercises subchunk-grid propagation")
-        }
-    }
-
-    fn local_only_grid_chain(expose_subchunks: bool, reject_chunk_local: bool) -> CodecChainBound {
-        let data_type = data_type::uint8();
-        let fill_value = FillValue::from(0u8);
-        let mut array_to_array: Vec<Arc<dyn ArrayToArrayCodecTraits>> =
-            vec![Arc::new(LocalOnlyReshapeGridCodec {
-                data_type: data_type.clone(),
-                fill_value: fill_value.clone(),
-                reject_chunk_local: false,
-            })];
-        if reject_chunk_local {
-            array_to_array.push(Arc::new(LocalOnlyReshapeGridCodec {
-                data_type: data_type.clone(),
-                fill_value: fill_value.clone(),
-                reject_chunk_local: true,
-            }));
-        }
-        CodecChainBound {
-            array_to_array,
-            array_to_bytes: Arc::new(TestSubchunkingCodec {
-                data_type,
-                fill_value,
-                expose_subchunks,
-            }),
-            bytes_to_bytes: vec![],
-            cache_index: None,
-        }
-    }
-
-    #[test]
-    fn codec_chain_discovers_chunk_local_grid_after_global_mapping_fails() {
-        let decoded_chunk_grid = ChunkGrid::new(
-            RegularChunkGrid::new(
-                vec![4, 6],
-                vec![NonZeroU64::new(2).unwrap(), NonZeroU64::new(6).unwrap()],
-            )
-            .unwrap(),
-        );
-        let chain = local_only_grid_chain(true, false);
-        assert!(matches!(
-            chain
-                .decoded_subchunk_grid((&decoded_chunk_grid).into())
-                .unwrap(),
-            ChunkGridDecoded::ChunkLocal
-        ));
-    }
-
-    #[test]
-    fn codec_chain_does_not_infer_chunk_local_without_downstream_subchunks() {
-        let decoded_chunk_grid = ChunkGrid::new(
-            RegularChunkGrid::new(
-                vec![4, 6],
-                vec![NonZeroU64::new(2).unwrap(), NonZeroU64::new(6).unwrap()],
-            )
-            .unwrap(),
-        );
-        let chain = local_only_grid_chain(false, false);
-        assert!(matches!(
-            chain
-                .decoded_subchunk_grid((&decoded_chunk_grid).into())
-                .unwrap(),
-            ChunkGridDecoded::None
-        ));
-    }
-
-    #[test]
-    fn codec_chain_allows_later_array_codec_to_reject_chunk_local() {
-        let decoded_chunk_grid = ChunkGrid::new(
-            RegularChunkGrid::new(
-                vec![4, 6],
-                vec![NonZeroU64::new(2).unwrap(), NonZeroU64::new(6).unwrap()],
-            )
-            .unwrap(),
-        );
-        let chain = local_only_grid_chain(true, true);
-        assert!(matches!(
-            chain
-                .decoded_subchunk_grid((&decoded_chunk_grid).into())
-                .unwrap(),
-            ChunkGridDecoded::None
-        ));
     }
 
     #[cfg(feature = "crc32c")]
